+++ game/unit_heat_seeker.cpp
@@ -22,7 +22,7 @@ void mark::unit::heat_seeker::tick(mark::tick_context& context) {
 		const auto turn_direction = mark::sgn(mark::atan(mark::rotate(direction, -m_rotation)));
 		const auto rot_step = static_cast<float>(turn_direction  * 500.f * dt);
 		if (std::abs(mark::atan(mark::rotate(direction, -m_rotation))) < 32.f * dt) {
-			m_rotation = mark::atan(direction);
+			m_rotation = static_cast<float>(mark::atan(direction));
 		} else {
 			m_rotation += rot_step;
 		}
